[script.openweathermap.maps] 1.0.4
