+++ script.openweathermap.maps/addon.xml
@@ -1,25 +1,23 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="script.openweathermap.maps" name="OpenWeatherMap Maps" version="1.0.3" provider-name="ronie">
+<addon id="script.openweathermap.maps" name="OpenWeatherMap Maps" version="1.0.4" provider-name="ronie">
 	<requires>
 		<import addon="xbmc.python" version="3.0.0"/>
 		<import addon="script.module.pil" version="1.1.7"/>
-		<import addon="script.module.requests" version="2.18.4"/>
+		<import addon="script.module.requests" version="2.22.0+matrix.1"/>
 	</requires>
 	<extension point="xbmc.python.library" library="default.py"/>
 	<extension point="xbmc.addon.metadata">
 		<summary lang="en_GB">Weather maps from openweathermap.org</summary>
 		<description lang="en_GB">Weather maps provided by OpenWeatherMap (http://openweathermap.org/)</description>
-		<language></language>
 		<platform>all</platform>
-		<license>GNU GENERAL PUBLIC LICENSE Version 2</license>
-		<forum></forum>
+		<license>GPL-2.0-only</license>
 		<website>https://forum.kodi.tv/showthread.php?tid=207110</website>
-		<email></email>
 		<source>https://gitlab.com/ronie/weather.openweathermap.maps/</source>
 		<assets>
 			<icon>resources/icon.png</icon>
 			<fanart>resources/fanart.jpg</fanart>
 		</assets>
+		<news>- code refactor</news>
 	</extension>
 </addon>
