+++ script.openweathermap.maps/changelog.txt
@@ -1,3 +1,6 @@
+v1.0.4
+- update
+
 v1.0.3
 - less image processing
