+++ script.openweathermap.maps/lib/utils.py
